+++ packages/splat-transform-native/source/include/node/api_buffer.h
@@ -1,6 +1,7 @@
 #pragma once
 #include <memory>
 #include <napi.h>
+#include <span>
 #include <vector>
 
 namespace node_api::buffer {