+++ packages/splat-transform-native/source/src/node/api_splat.cpp
@@ -32,10 +32,10 @@ enum SplatTable {
     SPLAT_TABLE_SH_OFFSET = 14
 };
 
-inline splat::Splat read_splat(const std::vector<Napi::Buffer<float>>& buffers, size_t sh_size, threading::ThreadPool& pool) {
+inline splat::Splat read_splat(const std::vector<std::span<float>>& buffers, size_t sh_size, threading::ThreadPool& pool) {
     auto thread_count = pool.thread_count();
     auto read_count = std::atomic<size_t>(0);
-    auto gaussian_count = buffers[0].Length();
+    auto gaussian_count = buffers[0].size();
     auto gaussians = std::vector<splat::Gaussian>(gaussian_count);
 
     auto gaussian_per_thread = std::max(gaussian_count / thread_count, static_cast<size_t>(1));
@@ -51,27 +51,27 @@ inline splat::Splat read_splat(const std::vector<Napi::Buffer<float>>& buffers, 
 
             auto& gaussian = gaussians[index];
             gaussian.mean = Eigen::Vector3f(
-                buffers[SPLAT_TABLE_MEAN_X].Data()[index],
-                buffers[SPLAT_TABLE_MEAN_Y].Data()[index],
-                buffers[SPLAT_TABLE_MEAN_Z].Data()[index]);
+                buffers[SPLAT_TABLE_MEAN_X][index],
+                buffers[SPLAT_TABLE_MEAN_Y][index],
+                buffers[SPLAT_TABLE_MEAN_Z][index]);
             gaussian.scale = Eigen::Vector3f(
-                buffers[SPLAT_TABLE_SCALE_X].Data()[index],
-                buffers[SPLAT_TABLE_SCALE_Y].Data()[index],
-                buffers[SPLAT_TABLE_SCALE_Z].Data()[index]);
+                buffers[SPLAT_TABLE_SCALE_X][index],
+                buffers[SPLAT_TABLE_SCALE_Y][index],
+                buffers[SPLAT_TABLE_SCALE_Z][index]);
             gaussian.rotation = Eigen::Vector4f(
-                buffers[SPLAT_TABLE_QUAT_X].Data()[index],
-                buffers[SPLAT_TABLE_QUAT_Y].Data()[index],
-                buffers[SPLAT_TABLE_QUAT_Z].Data()[index],
-                buffers[SPLAT_TABLE_QUAT_W].Data()[index]);
-            gaussian.opacity = buffers[SPLAT_TABLE_OPACITY].Data()[index];
+                buffers[SPLAT_TABLE_QUAT_X][index],
+                buffers[SPLAT_TABLE_QUAT_Y][index],
+                buffers[SPLAT_TABLE_QUAT_Z][index],
+                buffers[SPLAT_TABLE_QUAT_W][index]);
+            gaussian.opacity = buffers[SPLAT_TABLE_OPACITY][index];
 
             // SH
             gaussian.sh = ::splat::SH(sh_size + 3);
-            gaussian.sh[0] = buffers[SPLAT_TABLE_COLOR_R].Data()[index];
-            gaussian.sh[1] = buffers[SPLAT_TABLE_COLOR_G].Data()[index];
-            gaussian.sh[2] = buffers[SPLAT_TABLE_COLOR_B].Data()[index];
+            gaussian.sh[0] = buffers[SPLAT_TABLE_COLOR_R][index];
+            gaussian.sh[1] = buffers[SPLAT_TABLE_COLOR_G][index];
+            gaussian.sh[2] = buffers[SPLAT_TABLE_COLOR_B][index];
             for (auto j = 3; j < gaussian.sh.size(); j++) {
-                gaussian.sh[j] = buffers[SPLAT_TABLE_SH_OFFSET + j - 3].Data()[index];
+                gaussian.sh[j] = buffers[SPLAT_TABLE_SH_OFFSET + j - 3][index];
             }
 
             gaussian.rotation.normalize();
@@ -101,49 +101,54 @@ inline splat::Splat read_splat(const std::vector<Napi::Buffer<float>>& buffers, 
     return splat;
 }
 
-inline void write_splat(
+inline std::vector<Napi::Float32Array> write_splat(
+    Napi::Env& env,
+    size_t buffer_count,
     const splat::Splat& splat,
-    threading::ThreadPool& pool,
-    /* out */ std::span<std::unique_ptr<std::vector<float>>> buffers) {
+    threading::ThreadPool& pool) {
+    auto buffers = std::vector<Napi::Float32Array>();
+    auto buffer_spans = std::vector<std::span<float>>();
     auto thread_count = pool.thread_count();
     auto written_count = std::atomic<size_t>(0);
-    auto write_offset = buffers[0]->size();
+    buffers.reserve(buffer_count);
+    buffer_spans.reserve(buffer_count);
+
+    for (auto i = 0; i < buffer_count; i++) {
+        buffers.push_back(Napi::Float32Array::New(env, splat.gaussians.size()));
+        buffer_spans.emplace_back(buffers.back().Data(), splat.gaussians.size());
+    }
 
     auto gaussian_per_thread = std::max(splat.gaussians.size() / thread_count, static_cast<size_t>(1));
     auto futures = std::vector<std::future<void>>();
     auto used_threads = std::min(splat.gaussians.size(), thread_count);
 
-    // prepare writable range.
-    for (auto& buffer : buffers) {
-        buffer->resize(buffer->size() + splat.gaussians.size());
-    }
     futures.reserve(used_threads);
     written_count.store(0, std::memory_order_release);
 
     // transform gaussian to js readable struct
-    auto process_gaussian = [&, write_offset](size_t start, size_t count) {
+    auto process_gaussian = [&](size_t start, size_t count) {
         for (auto i = 0; i < count; i++) {
             auto index = start + i;
-            auto write_index = write_offset + index;
+            auto write_index = index;
             auto& gaussian = splat.gaussians[index];
 
-            (*buffers[SPLAT_TABLE_MEAN_X])[write_offset + index] = gaussian.mean.x();
-            (*buffers[SPLAT_TABLE_MEAN_Y])[write_offset + index] = gaussian.mean.y();
-            (*buffers[SPLAT_TABLE_MEAN_Z])[write_offset + index] = gaussian.mean.z();
-            (*buffers[SPLAT_TABLE_SCALE_X])[write_offset + index] = gaussian.scale.x();
-            (*buffers[SPLAT_TABLE_SCALE_Y])[write_offset + index] = gaussian.scale.y();
-            (*buffers[SPLAT_TABLE_SCALE_Z])[write_offset + index] = gaussian.scale.z();
-            (*buffers[SPLAT_TABLE_QUAT_X])[write_offset + index] = gaussian.rotation.x();
-            (*buffers[SPLAT_TABLE_QUAT_Y])[write_offset + index] = gaussian.rotation.y();
-            (*buffers[SPLAT_TABLE_QUAT_Z])[write_offset + index] = gaussian.rotation.z();
-            (*buffers[SPLAT_TABLE_QUAT_W])[write_offset + index] = gaussian.rotation.w();
-            (*buffers[SPLAT_TABLE_COLOR_R])[write_offset + index] = gaussian.sh[0];
-            (*buffers[SPLAT_TABLE_COLOR_G])[write_offset + index] = gaussian.sh[1];
-            (*buffers[SPLAT_TABLE_COLOR_B])[write_offset + index] = gaussian.sh[2];
-            (*buffers[SPLAT_TABLE_OPACITY])[write_offset + index] = gaussian.opacity;
+            buffer_spans[SPLAT_TABLE_MEAN_X][write_index] = gaussian.mean.x();
+            buffer_spans[SPLAT_TABLE_MEAN_Y][write_index] = gaussian.mean.y();
+            buffer_spans[SPLAT_TABLE_MEAN_Z][write_index] = gaussian.mean.z();
+            buffer_spans[SPLAT_TABLE_SCALE_X][write_index] = gaussian.scale.x();
+            buffer_spans[SPLAT_TABLE_SCALE_Y][write_index] = gaussian.scale.y();
+            buffer_spans[SPLAT_TABLE_SCALE_Z][write_index] = gaussian.scale.z();
+            buffer_spans[SPLAT_TABLE_QUAT_X][write_index] = gaussian.rotation.x();
+            buffer_spans[SPLAT_TABLE_QUAT_Y][write_index] = gaussian.rotation.y();
+            buffer_spans[SPLAT_TABLE_QUAT_Z][write_index] = gaussian.rotation.z();
+            buffer_spans[SPLAT_TABLE_QUAT_W][write_index] = gaussian.rotation.w();
+            buffer_spans[SPLAT_TABLE_COLOR_R][write_index] = gaussian.sh[0];
+            buffer_spans[SPLAT_TABLE_COLOR_G][write_index] = gaussian.sh[1];
+            buffer_spans[SPLAT_TABLE_COLOR_B][write_index] = gaussian.sh[2];
+            buffer_spans[SPLAT_TABLE_OPACITY][write_index] = gaussian.opacity;
 
             for (auto j = 3; j < gaussian.sh.size(); j++) {
-                (*buffers[SPLAT_TABLE_SH_OFFSET + j - 3])[write_offset + index] = gaussian.sh[j];
+                buffer_spans[SPLAT_TABLE_SH_OFFSET + j - 3][write_index] = gaussian.sh[j];
             }
         }
 
@@ -161,6 +166,8 @@ inline void write_splat(
     }
 
     assert(written_count.load(std::memory_order_acquire) == splat.gaussians.size());
+
+    return buffers;
 }
 } // namespace
 namespace node_api::splat {
@@ -178,11 +185,12 @@ Napi::Value generate_lod(const Napi::CallbackInfo& info) {
     std::vector<::splat::Splat> blocks;
     // read & create blocks.
     {
-        auto buffers = std::vector<Napi::Buffer<float>>();
+        auto buffers = std::vector<std::span<float>>();
         auto array = info[0].As<Napi::Array>();
         buffers.reserve(sh_size + SPLAT_TABLE_SH_OFFSET);
         for (auto i = 0; i < sh_size + SPLAT_TABLE_SH_OFFSET; i++) {
-            buffers.push_back(array[i].AsValue().As<Napi::Buffer<float>>());
+            auto buffer = array[i].AsValue().As<Napi::Buffer<float>>();
+            buffers.emplace_back(buffer.Data(), buffer.Length());
         }
         blocks = ::splat::block::split(
             read_splat(buffers, sh_size, pool),
@@ -190,6 +198,7 @@ Napi::Value generate_lod(const Napi::CallbackInfo& info) {
     }
 
     auto level_parameters = info[2].As<Napi::Buffer<::splat::lod::SplatLevelParameters>>();
+    auto levels = level_parameters.Length();
     auto level_parameters_span = std::span(level_parameters.Data() + 1, level_parameters.Length() - 1);
 
     auto results = std::vector<::splat::Splat>();
@@ -198,13 +207,12 @@ Napi::Value generate_lod(const Napi::CallbackInfo& info) {
     auto block_refs = std::make_unique<std::vector<uint32_t>>();
 
     // reverse data, small.
-    results.reserve(blocks.size() * level_parameters.Length());
-    gaussian_count->reserve(blocks.size() * level_parameters.Length());
+    results.reserve(blocks.size() * levels);
+    gaussian_count->reserve(blocks.size() * levels);
     block_boxes->reserve(blocks.size() * 6);
-    block_refs->reserve(blocks.size() * level_parameters.Length());
+    block_refs->reserve(blocks.size() * levels);
 
     {
-        auto levels = level_parameters.Length();
         auto futures = std::vector<std::future<::splat::lod::SplatLod>>();
         auto used_threads = std::min(blocks.size(), thread_count);
         auto min_size = info[4].As<Napi::Number>().Uint32Value();
@@ -215,8 +223,8 @@ Napi::Value generate_lod(const Napi::CallbackInfo& info) {
         auto process_block = [&, levels, max_step](size_t index) -> ::splat::lod::SplatLod {
             auto& block = blocks[index];
             auto lod = ::splat::lod::SplatLod();
-            lod.levels.reserve(level_parameters.Length());
-            lod.splats.reserve(level_parameters.Length());
+            lod.levels.reserve(levels);
+            lod.splats.reserve(levels);
             lod.levels.push_back(0);
             lod.splats.push_back(std::move(block));
             if (level_parameters_span.size() > 0) {
@@ -256,18 +264,14 @@ Napi::Value generate_lod(const Napi::CallbackInfo& info) {
 
     // auto total = std::reduce(gaussian_count->begin(), gaussian_count->end());
     auto buffers_per_splat = sh_size + SPLAT_TABLE_SH_OFFSET;
-    auto buffers = std::vector<std::unique_ptr<std::vector<float>>>();
+    auto buffers = std::vector<Napi::Float32Array>();
 
     buffers.reserve(buffers_per_splat * results.size());
 
     for (auto& splat : results) {
-        // prealloc buffer before insertion.
-        for (auto i = 0; i < buffers_per_splat; i++) {
-            buffers.push_back(std::make_unique<std::vector<float>>());
-            buffers.back()->reserve(splat.gaussians.size());
-        }
-
-        write_splat(splat, pool, std::span(buffers.begin() + buffers.size() - buffers_per_splat, buffers_per_splat));
+        helpers::container::append_range(
+            buffers,
+            write_splat(env, buffers_per_splat, splat, pool));
 
         // free data already transformed.
         {
@@ -281,7 +285,7 @@ Napi::Value generate_lod(const Napi::CallbackInfo& info) {
         auto offset = 0;
         object.Set("data", data);
         for (auto i = 0; i < buffers.size(); i++) {
-            data[i] = node_api::buffer::UniqueVecBufferFinalizer<float>::make_buffer(env, std::move(buffers[i]));
+            data[i] = buffers[i];
         }
         auto _ = std::move(buffers);
     }
@@ -305,11 +309,12 @@ Napi::Value split(const Napi::CallbackInfo& info) {
     std::vector<::splat::Splat> blocks;
     // read & create blocks.
     {
-        auto buffers = std::vector<Napi::Buffer<float>>();
+        auto buffers = std::vector<std::span<float>>();
         auto array = info[0].As<Napi::Array>();
         buffers.reserve(sh_size + SPLAT_TABLE_SH_OFFSET);
         for (auto i = 0; i < sh_size + SPLAT_TABLE_SH_OFFSET; i++) {
-            buffers.push_back(array[i].AsValue().As<Napi::Buffer<float>>());
+            auto buffer = array[i].AsValue().As<Napi::Buffer<float>>();
+            buffers.emplace_back(buffer.Data(), buffer.Length());
         }
         blocks = ::splat::block::split(
             read_splat(buffers, sh_size, pool),
@@ -319,20 +324,15 @@ Napi::Value split(const Napi::CallbackInfo& info) {
     auto buffers_per_splat = sh_size + SPLAT_TABLE_SH_OFFSET;
     auto block_boxes = std::make_unique<std::vector<float>>();
     auto gaussian_count = std::make_unique<std::vector<uint32_t>>();
-    auto buffers = std::vector<std::unique_ptr<std::vector<float>>>();
+    auto buffers = std::vector<Napi::Float32Array>();
 
     block_boxes->reserve(blocks.size() * 6);
     gaussian_count->reserve(blocks.size());
     buffers.reserve(buffers_per_splat * blocks.size());
 
     for (auto& splat : blocks) {
-        // prealloc buffer before insertion.
-        for (auto i = 0; i < buffers_per_splat; i++) {
-            buffers.push_back(std::make_unique<std::vector<float>>());
-            buffers.back()->reserve(splat.gaussians.size());
-        }
 
-        write_splat(splat, pool, std::span(buffers.begin() + buffers.size() - buffers_per_splat, buffers_per_splat));
+        helpers::container::append_range(buffers, write_splat(env, buffers_per_splat, splat, pool));
         gaussian_count->push_back(static_cast<uint32_t>(splat.gaussians.size()));
         {
             auto& bbx = splat.bounding_box;
@@ -352,7 +352,7 @@ Napi::Value split(const Napi::CallbackInfo& info) {
         auto offset = 0;
         object.Set("data", data);
         for (auto i = 0; i < buffers.size(); i++) {
-            data[i] = node_api::buffer::UniqueVecBufferFinalizer<float>::make_buffer(env, std::move(buffers[i]));
+            data[i] = buffers[i];
         }
         auto _ = std::move(buffers);
     }